clippy

+++ disarm64-cli/src/main.rs
@@ -384,7 +384,7 @@ fn decode_pe(data: &[u8], benchmark: Option<BenchmarkMode>) -> anyhow::Result<Pr
     let mut buffer = String::new();
     for section in &pe.sections {
         if section.characteristics & 0x20000000 != 0 {
-            let vbase = pe.image_base + section.virtual_address as usize;
+            let vbase = pe.image_base + section.virtual_address as u64;
             log::info!(
                 "// Decoding section {:?} @ {vbase:#x}",
                 std::ffi::CStr::from_bytes_until_nul(&section.name).unwrap_or(
@@ -394,7 +394,7 @@ fn decode_pe(data: &[u8], benchmark: Option<BenchmarkMode>) -> anyhow::Result<Pr
 
             let data =
                 &data[section.pointer_to_raw_data as usize..][..section.size_of_raw_data as usize];
-            let stats = process_bytes(data, vbase as u64, &mut buffer, benchmark)?;
+            let stats = process_bytes(data, vbase, &mut buffer, benchmark)?;
             proc_stats.add(&stats);
         }
     }
